assignment editor tweaks
